Day11_assignment

+++ app/Http/Controllers/Users/UserController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers\Users;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Rules\Password;
 use Illuminate\Validation\Rules;
 use Illuminate\Validation\Rules\Password as RulesPassword;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
@@ -17,7 +19,10 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
+        if(!Gate::allows('user_list')){
+            abort(401);
+        }
+        $users = User::with('roles')->get();
         return view('users.index',compact('users'));
     }
 
@@ -27,7 +32,11 @@ class UserController extends Controller
     public function create()
     {
         //
-        return view('users.create');
+        if(!Gate::allows('user_create')){
+            abort(401);
+        }
+        $roles = Role::all();
+        return view('users.create',compact('roles'));
     }
 
     /**
@@ -37,17 +46,22 @@ class UserController extends Controller
     public function store(Request $request)
     {
         //
+        if(!Gate::allows('user_store')){
+            abort(401);
+        }
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
             'password' => ['required', 'confirmed', RulesPassword::defaults()],
+            'role_id' => ['required','integer','min:1']
         ]);
-
+        $role = Role::where('id',$request->role_id)->first();
         $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
             'password' => Hash::make($request->password),
         ]);
+        $user->assignRole($role);
         return redirect()->route('users.index');
     }
 
@@ -65,6 +79,9 @@ class UserController extends Controller
     public function edit(string $id)
     {
         //
+        if(!Gate::allows('user_edit')){
+            abort(401);
+        }
         $user = User::where('id',$id)->first();
         return view('users.edit',compact('user'));
     }
@@ -82,6 +99,9 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!Gate::allows('user_destory')){
+            abort(401);
+        }
         User::where('id',$id)->delete();
         // return view('users.index');
         return redirect()->route('users.index');
